US79 -Discovered missing constraint on EntityMapper where quickanswer field was not required (it should be).
In the process of adding tests, discovered that uniqueness validation was being
done in the controller.save and .update methods instead of in a custom
validator in the domain object.  I believe this was because someone remember
there used to be an issue using queris in customer validator, but that seems to be
resolved in current versions of grails. Moved the validation  into the domain object