US25 - More yak shaving. Coalesce dataElementName and categories fields into a single filterString field in AffStatisticDescriptor
Like OpenApiStatisticDescriptors, we need to constrain AffStatisticDescriptor 
instances to be unique.  For AffStatisticDescriptor, instances must be unique
across the combination of data set, table and filters.  The filters
were previously broken into two components - dataElementName and
categories.  Since only one or the other of those two fields can be specified, both
had to be marked optional, and you cant enforce a uniqueness constraint on
optional fields.  So the first step in enforcing uniqueness is to coalesce those
two fields into a single, non-optional, filterString field.