US25 - Tighten up duplicate checking when creating or editing AffStatisticDescriptors
If editing, locatorId indicates the descriptor being edited.  If creating new, locatorId
should be null (or not specified).  Also, if editing (locatorId is specified), but no
descriptor exists with that id, then a new descriptor will be created.  The only time
this can happen is if some one deletes the descriptor after the user has started
editing, but before they clicked save. The thinking is that if the user is hitting
the save button on the edit screen, they want the descriptor to exist, so we 
should just go ahead an create it rather than giving them an error and making
them go through the create process from scratch.