US25 - Fill in Display Name from topic info for aff statistic descriptor
Also updated the api provider selection on the create to match the style
of the rest of the create form.  Removed ability to select api provider - we've
agreed that for now, allowing the api to be changed makes things a lot
more complicated than we want to deal with right now.