US127, Jira SP-784 - Move statistic delete to service to force transaction.
Statistic deletion needs to be wrapped in a transaction because there is a cascade
delete to clean up synonyms associated with the statistic. 
In order to test transactional behavior, we had to turn off transactions in the
integration test (nested transactions dont work the way we need them too). 
As a result, writes to the database during the integration test are not rolled back,
and other integration tests can no longer assume a clean database when they
run. Had to fix a few tests to account for this.