US127, Jira SP-795 - Fix column sorting
At some point, the DataTables plugin was added to the project. It handles 
column sorting and pagination on the client. This conflicts with grails' 
sortableColumn tag which uses server-side sorting. This fix removes all of the 
sortableColumn tags except the ones for the counties and places tables which 
require server-side sorting due to the amount of data in those two tables.