US127, Jira SP-796 - Fix pagination
At some point, the DataTables plugin was added to the project. It handles 
column sorting and pagination on the client. This conflicts with grails' 
default pagination which is done server-side. This fix modifies all of the 
controllers to always return all entities in the list() method (which is what 
DataTables expects), rather than just a page.