US129 - Add IntelliJ code analysis profile along with processing tool.
Add the IntelliJ code analysis profile that census uses when they review our code,
inluding comments on how to run the analysis with the profile.

Created a simple groovy script to process the output of the IntelliJ analsysis and
produce csv output that will make it easier to review the problems identified
by the analysis tool.