US127, Jira SP-821 - Invalid apikey parameter cause infinite exception loop
Corrected the 'if' statement to avoid the NPE, and cleaned up the code a little.
Also put a guard at the beginning of the filter to check if the error controller is
being processed, and if it is, just exit without doing anything.  There is supposed
to be a way to have the filter exclude the error controller, but I was not able to get
that to work