US153 - Log user activity on SPA - fix issue with AuditLogger in unit tests
Failed to notice an issue with using the AuditLogger in unit test environment.
AuditLogger uses new User().domainClass to get a reference to the spring security
service.  However, the domainClass propery is not setup in the unit test 
environment.  Moved ResultsExtractorTest to integration tests (that's where it
should have been in the first place).  Deleted UimaControllerTest (it became
irrelevant once ResultsExtractor and its corresponding tests were added.  
StatisticServiceTests was the only other issue, and that was fixed by injecting
a mock AuditLogger into the UimaService.