SP-818 - This was fixed in enhanced search.  Applying same to ATO:
The resources plugin will redirect any requests for items under web-app to
/static, so js/*.js gets redirected to static/js/*.js, etc. The spring security plugin was only configure to allow anonymous access to css/*, js/*, images/*, etc, so the initial request would pass security, but then the redirect to /static would fail security. Since the login page has references to js and css, it would not display properly.

As to why this worked before (and still works on local deployments), my only guess is that it is due to the order in which grails plugins are loaded (the load order is undefined). Thus, it may be that if the resources plugin gets loaded first, it will satisfy all requests for static content before the
security plugin even sees the request.